requirement.txt
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.2
Brotli @ file:///private/tmp/python-brotli-20231117-5377-1dmlxkj/Brotli-1.1.0
cattrs==23.2.3
certifi==2022.12.7
charset-normalizer==3.1.0
empyrical==0.5.5
frozendict==2.4.0
helium==3.0.9
html5lib==1.1
idna==3.4
lxml==5.1.0
markovify==0.9.4
MarkupSafe @ file:///private/tmp/python-markupsafe-20240204-4734-rcyty/MarkupSafe-2.1.5
multitasking==0.0.11
mutagen @ file:///private/tmp/python-mutagen-20231211-52937-mx9s2c/mutagen-1.47.0
numpy==1.24.3
packaging==23.0
pandas==2.2.1
pandas-datareader==0.10.0
pandas-finance==0.1.3
peewee==3.17.1
platformdirs==4.2.0
Pygments @ file:///private/tmp/pygments-20231121-5274-1rlyumb/pygments-2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
scipy==1.12.0
selenium==3.141.0
six==1.16.0
soupsieve==2.4
tqdm==4.65.0
tzdata==2024.1
Unidecode==1.3.6
url-normalize==1.4.3
urllib3==1.26.15
webdriver-manager==3.8.5
webencodings==0.5.1
yfinance==0.2.37
